pyproject.toml
[tool.poetry]
name = "drf-docker-postgres"
version = "0.1.0"
description = "CURD operation with django rest framework with postgres and docker"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django-admin-cli = "^0.1.1"
django = "^5.1.1"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.9"
drf-yasg = "^1.21.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
